requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
razorpay==1.4.2
reportlab==4.2.2
chardet==5.2.0
